Break block processing into separate files
